+++ scripts/CLDL/include/cldl/Neuron.h
@@ -15,40 +15,40 @@ class Neuron {
 public:
     Neuron(int _nInputs, int _nrawInternalErrors);
     ~Neuron();
-    double getRawInternalErrors(int _internalErrorIndex);
+    float getRawInternalErrors(int _internalErrorIndex);
     enum biasInitMethod {B_NONE = 0, B_RANDOM = 1};
     enum weightInitMethod {W_ZEROS = 0, W_ONES = 1, W_RANDOM = 2};
     enum actMethod {Act_Sigmoid = 0, Act_Tanh = 1, Act_NONE = 2};
     enum errorMethod {Value = 1, Absolute = 2, Sign = 3};
     void initNeuron(int _neuronIndex, int _layerIndex, weightInitMethod _wim,
                     biasInitMethod _bim, actMethod _am);
-    void setLearningRate(double _learningRate);
-    void setInput(int _index, double _value);
-    void propInputs(int _index, double _value);
+    void setLearningRate(float _learningRate);
+    void setInput(int _index, float _value);
+    void propInputs(int _index, float _value);
     int calcOutput(int _layerHasReported);
-    void setInternalError(int _internalErrorIndex, double _sumValue,
+    void setInternalError(int _internalErrorIndex, float _sumValue,
                           errorMethod _errorMethod);
     void setErrorInputsAndCalculateInternalError(int _inputIndex,
-                                                 double _value, int _internalErrorIndex,
+                                                 float _value, int _internalErrorIndex,
                                                  errorMethod _errorMethod);
     void updateWeights();
-    double doActivation(double _sum);
-    double doActivationPrime(double _input);
-    double getOutput();
-    double getSumOutput();
-    double getWeights(int _inputIndex);
-    double getInputs(int _inputIndex);
-    double getInitWeights(int _inputIndex);
-    double getWeightChange();
-    double getMaxWeight();
-    double getMinWeight();
-    double getSumWeight();
-    double getWeightDistance();
+    float doActivation(float _sum);
+    float doActivationPrime(float _input);
+    float getOutput();
+    float getSumOutput();
+    float getWeights(int _inputIndex);
+    float getInputs(int _inputIndex);
+    float getInitWeights(int _inputIndex);
+    float getWeightChange();
+    float getMaxWeight();
+    float getMinWeight();
+    float getSumWeight();
+    float getWeightDistance();
     int getnInputs();
 
     void saveWeights();
     void printNeuron();
-    inline void setWeight(int _index, double _weight) {
+    inline void setWeight(int _index, float _weight) {
         assert((_index >= 0) && (_index < nInputs));
         weights[_index] = _weight;
     }
@@ -56,31 +56,31 @@ public:
 private:
     // initialisation:
     int nInputs = 0;
-    double weightBoost = 1000;
+    float weightBoost = 1000;
     int myLayerIndex = 0;
     int myNeuronIndex = 0;
-    double *initialWeights = nullptr;
-    double learningRate = 0;
-    double *inputErrors = nullptr;
+    float *initialWeights = nullptr;
+    float learningRate = 0;
+    float *inputErrors = nullptr;
     int numBuses = 0;
-    double *rawInternalErrors = nullptr;
-    double* internalErrors = nullptr;
+    float *rawInternalErrors = nullptr;
+    float* internalErrors = nullptr;
     bool *busIsSet = nullptr;
     int countInputErrors = 0;
     int* busMethod = nullptr;
-    double resultantInternalError = 1;
+    float resultantInternalError = 1;
     
     int iHaveReported = 0;
-    double *inputs = nullptr;
-    double bias = 0;
-    double sum = 0;
-    double output = 0;
+    float *inputs = nullptr;
+    float bias = 0;
+    float sum = 0;
+    float output = 0;
 
-    double *weights = nullptr;
-    double weightSum = 0;
-    double maxWeight = 1;
-    double minWeight = 1;
-    double weightChange = 0;
-    double weightsDifference = 0;
+    float *weights = nullptr;
+    float weightSum = 0;
+    float maxWeight = 1;
+    float minWeight = 1;
+    float weightChange = 0;
+    float weightsDifference = 0;
     int actMet = 0;
 };
